feat(indexer): add RegistryEntity and split handlers into separate files
- Add RegistryEntity to track live registry state (owner, registryFeeShare)
  bootstrapped from the OwnershipTransferred event emitted by the Ownable
  constructor — no dedicated RegistryCreated event exists on-chain
- Add registryId FK and link AssetEntity → RegistryEntity via relations
- Add subscriptionPrice and tokenAddress to AssetEntity; subscriptionPrice
  is kept current by the SubscriptionPriceUpdated handler
- Split src/index.ts into src/handlers/registry.ts, src/handlers/asset.ts,
  and src/utils.ts for shared ID helpers
